Add tasks for mt5
This adds a few tasks that mt5 can do. These tasks, such as Glue, are already
present in the registry from t5, both those tasks are stuck using the t5
sentencepiece vocabulary. These versions of the tasks, prepended with `mt5_`
provide version of the task that use the mt5 sentence piece vocab.

PiperOrigin-RevId: 357203264